refactor(orchestrator): split orchestrator.go into finalize.go + resourceset.go
Same per-concern split pattern as #253 applied to the orchestrator
package. The 815 LOC orchestrator.go mixed four concerns:

  - lifecycle: Config, Orchestrator struct, Result, New, Bootstrap,
    Run, Render, Stop
  - finalize/reporting: detectOrphans, demoteOrphans, finalize,
    logSummary, logResourceFailures, aggregateFailures
  - ResourceSet post-run: expandResourceSetsPostRun, dedupKeyForDoc,
    resolveInputProvider

Sibling files cycles.go (cycle detection) and rendered.go (parent-
provenance tracker) already followed this pattern; this commit
brings the remaining clusters out of the monolith.

New layout:

  pkg/orchestrator/
    orchestrator.go  (Config, Orchestrator, Result, New, lifecycle, validateDependsOn, buildChangeFilter, Render)
    finalize.go      (detectOrphans + demoteOrphans + finalize + log*)
    resourceset.go   (expandResourceSetsPostRun + helpers)
    cycles.go        (unchanged)
    rendered.go      (unchanged)

Net: orchestrator.go drops 815 → 504 LOC, focused on the reconcile
lifecycle. No behavior change; identical imports re-distributed.

Verified:
- go test -count=1 ./...      — full suite green
- golangci-lint run ./...     — 0 issues
- tholinka/home-ops           — 118/0
- JJGadgets/Biohazard         — 199/0
- m00nwtchr/homelab-cluster   — 287/2 (pre-existing 404)
- buroa/k8s-gitops            — 73/0
- Zariel/home-ops             — 93/0
